requirements.txt
logging==0.4.9.6
RPi.GPIO==0.5.3a

